Move SyncHelper to library
